壁近でジャンプした後に動けなくなるバグを修正。 ApplyDamageの引数をfloat→intに変更。

+++ Client/Assets/99_User/04_Nakamoto/02_Script/Player/Player.cs
@@ -18,5 +18,5 @@ abstract public class Player : MonoBehaviour
     /// </summary>
     /// <param name="damage">ダメージ量</param>
     /// <param name="position">攻撃したオブジェの位置</param>
-    abstract public void ApplyDamage(float damage, Vector3 position);
+    abstract public void ApplyDamage(int damage, Vector3 position);
 }
